+++ lib/fluent_result.dart
@@ -1,6 +1,7 @@
 library fluent_result;
 
+export 'src/errors/errors.dart';
 export 'src/extensions.dart';
+export 'src/methods.dart';
 export 'src/result.dart';
-export 'src/result_error.dart';
 export 'src/result_of.dart';