+++ lib/src/errors/errors.dart
@@ -1,0 +1,2 @@
+export 'exception_error.dart';
+export 'result_error.dart';