+++ lib/src/errors/exception_error.dart
@@ -1,0 +1,10 @@
+import 'package:fluent_result/fluent_result.dart';
+
+/// Result Error caused by exception
+class ResultException extends ResultError {
+  ///
+  ResultException(this.exception) : super(exception.toString());
+
+  ///
+  final Exception exception;
+}