+++ lib/src/methods.dart
@@ -1,0 +1,16 @@
+import 'package:fluent_result/fluent_result.dart';
+
+/// Success result
+Result success() {
+  return Result.ok;
+}
+
+/// Success result with value
+ResultOf<T> successWith<T>(T value) {
+  return ResultOf.success(value);
+}
+
+/// Fail result with reason
+ResultOf<T?> fail<T>(Object reason) {
+  return ResultOf.withErr<T>(reason);
+}