+++ lib/src/result_of.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:fluent_result/fluent_result.dart';
 
 /// Generic version of `Result` that holds value
@@ -28,8 +30,7 @@ class ResultOf<T> extends Result {
   /// ```dart
   /// ResultOf.fail<MyObject>(ResultError('fail reason'));
   /// ```
-  @Deprecated('use withError instead')
-  static ResultOf<T?> fail<T>(ResultError? error) {
+  static ResultOf<T?> fail<T>(ResultError error) {
     return ResultOf<T?>(isSuccess: false, value: null, error: error);
   }
 
@@ -38,26 +39,61 @@ class ResultOf<T> extends Result {
   /// ```dart
   /// ResultOf.withError<MyObject>(ResultError('fail reason'));
   /// ```
-  static ResultOf<T?> withError<T>(ResultError error) {
-    return ResultOf<T?>(isSuccess: false, value: null, error: error);
+  @Deprecated('Use `withErr(object)`')
+  static ResultOf<T?> withError<T>(ResultError error) => withErr(error);
+
+  ///
+  @Deprecated('Use `withErr(object)`')
+  static ResultOf<T?> withErrorMessage<T>(String message) => withErr(message);
+
+  ///
+  @Deprecated('Use `withErr(object)`')
+  static ResultOf<T?> withException<T>(Exception exception) =>
+      withErr(exception);
+
+  ///
+  static ResultOf<T?> withErr<T>(Object object) {
+    if (object is Exception) {
+      return ResultOf.fail(ResultException(object));
+    }
+    if (object is Error) {
+      return ResultOf.fail(ResultError.of(object));
+    }
+    if (object is ResultError) {
+      return ResultOf.fail(object);
+    }
+    return ResultOf.fail(ResultError(object.toString()));
   }
 
-  ///
-  static ResultOf<T?> withErrorMessage<T>(String message) =>
-      ResultOf<T?>(isSuccess: false, value: null, error: ResultError(message));
+  /// Wrapped on try/catch
+  static ResultOf<T?> trySync<T>(T Function() func) {
+    try {
+      final data = func();
+      return ResultOf.success(data);
+    } catch (e) {
+      log(e.toString());
+      return ResultOf.withErr(e);
+    }
+  }
 
-  ///
-  static ResultOf<T?> withException<T>(Exception exception) => ResultOf<T?>(
-        isSuccess: false,
-        value: null,
-        error: ResultException(exception),
-      );
+  /// Wrapped on try/catch
+  static Future<ResultOf<T?>> tryAsync<T>(Future<T> Function() func) async {
+    try {
+      final data = await func();
+      return ResultOf.success(data);
+    } catch (e) {
+      log(e.toString());
+      return ResultOf.withErr(e);
+    }
+  }
 
   /// <summary>
   /// Convert result with value to result with another value. Use valueConverter
   /// parameter to specify the value transformation logic.
+  ///
+  /// No need valueConverter for Fail result. But for Success you should define it.
   /// </summary>
-  ResultOf<U?> toResult<U>({U Function(T)? valueConverter}) {
+  ResultOf<U?> map<U>([U Function(T)? valueConverter]) {
     if (isSuccess) {
       if (valueConverter == null) {
         throw Exception(
@@ -67,6 +103,6 @@ class ResultOf<T> extends Result {
       return ResultOf.success<U>(valueConverter(value!));
     }
 
-    return ResultOf.withError<U>(error!);
+    return fail(error!);
   }
 }