+++ test/convert_results_test.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: avoid_redundant_argument_values
 
 import 'package:fluent_result/fluent_result.dart';
+import 'package:shouldly/shouldly.dart';
 import 'package:test/test.dart';
 
 import 'helpers.dart';
@@ -12,8 +13,7 @@ void main() {
 
       expect(result.isSuccess, true);
 
-      final detailedResult =
-          result.toResult(valueConverter: (customer) => User(customer.id));
+      final detailedResult = result.map((customer) => User(customer.id));
       expect(detailedResult.value is User, true);
       expect(detailedResult.value!.id == 1, true);
     });
@@ -21,12 +21,56 @@ void main() {
 
   group('Convert fail Result', () {
     test('to another fail Result', () {
-      final result = ResultOf.withError<User>(ResultError('user not found'));
+      const reason = 'customer not found';
+      final res = getCustomer(failReason: reason);
+      res.isFail.should.beTrue();
+      res.errorMessage.should.be(reason);
+      res.error.should.beOfType<CustomerNotFound>();
+    });
 
-      expect(result.isFail, true);
-
-      final detailedResult = result.toResult<Customer>();
-      expect(detailedResult.value == null, true);
+    test('to another fail Result on user', () {
+      const reason = 'user not found';
+      final res = getCustomerWithFailOnUser(failReason: reason);
+      res.isFail.should.beTrue();
+      res.errorMessage.should.be(reason);
+      res.error.should.beOfType<UserNotFound>();
     });
   });
+}
+
+ResultOf<Customer?> getCustomerWithFailOnUser({String? failReason}) {
+  final userResult = getUser(failReason: failReason);
+  if (userResult.isFail) {
+    return userResult.map();
+  }
+
+  return Customer(1, 'Andrew').asResult();
+}
+
+ResultOf<Customer?> getCustomer({String? failReason}) {
+  final userResult = getUser();
+  if (userResult.isFail) {
+    return userResult.map();
+  }
+
+  if (failReason != null) {
+    return ResultOf.withErr(CustomerNotFound(failReason));
+  }
+
+  return Customer(1, 'Andrew').asResult();
+}
+
+ResultOf<User?> getUser({String? failReason}) {
+  if (failReason != null) {
+    return ResultOf.withErr(UserNotFound(failReason));
+  }
+  return const User(1).asResult();
+}
+
+class UserNotFound extends ResultError {
+  UserNotFound(String message) : super(message);
+}
+
+class CustomerNotFound extends ResultError {
+  CustomerNotFound(String message) : super(message);
 }