+++ test/fluent_result_test.dart
@@ -16,7 +16,7 @@ void main() {
 
     test('is fail and has message', () {
       const errorMessage = 'Some error';
-      final result = Result.withError(ResultError(errorMessage));
+      final result = Result.withErr(const ResultError(errorMessage));
       result.isFail.should.beTrue();
       result.isSuccess.should.beFalse();
       expect(result.errorMessage, errorMessage);
@@ -72,6 +72,90 @@ void main() {
       expect(result.isFail, true);
       expect(result.error is ResultException, true);
       expect(result.error!.message.contains(message), true);
+    });
+  });
+
+  group('wrapper', () {
+    test('success sync call', () {
+      final res = Result.trySync(() {
+        print('Done');
+      });
+
+      res.isSuccess.should.beTrue();
+    });
+
+    test('fail sync call', () {
+      final res = Result.trySync(() {
+        throw 'Some exception';
+      });
+
+      res.isFail.should.beTrue();
+      res.errorMessage.should.not.beNullOrEmpty();
+    });
+
+    test('success async call', () async {
+      final res = await Result.tryAsync(() async {
+        await Future.delayed(const Duration(seconds: 2));
+        print('Done');
+      });
+
+      res.isSuccess.should.beTrue();
+    });
+
+    test('fail async call', () async {
+      const errMessage = 'Some exception';
+      final res = await Result.tryAsync(() async {
+        await Future.delayed(const Duration(seconds: 2));
+        print('Done');
+        throw errMessage;
+      });
+
+      res.isFail.should.beTrue();
+      res.errorMessage.should.be(errMessage);
+    });
+  });
+
+  group('wrapper of', () {
+    test('success sync call', () {
+      final res = ResultOf.trySync(() {
+        print('Done');
+        return 1;
+      });
+
+      res.isSuccess.should.beTrue();
+      res.value.should.be(1);
+    });
+
+    test('fail sync call', () {
+      final res = ResultOf.trySync(() {
+        throw 'Some exception';
+      });
+
+      res.isFail.should.beTrue();
+      res.errorMessage.should.not.beNullOrEmpty();
+    });
+
+    test('success async call', () async {
+      final res = await ResultOf.tryAsync(() async {
+        await Future.delayed(const Duration(seconds: 2));
+        print('Done');
+        return 2;
+      });
+
+      res.isSuccess.should.beTrue();
+      res.value.should.be(2);
+    });
+
+    test('fail async call', () async {
+      const errMessage = 'Some exception';
+      final res = await ResultOf.tryAsync<bool>(() async {
+        await Future.delayed(const Duration(seconds: 2));
+        print('Done');
+        throw errMessage;
+      });
+
+      res.isFail.should.beTrue();
+      res.errorMessage.should.be(errMessage);
     });
   });
 }